+++ src/content/index.ts
@@ -1,5 +1,6 @@
 import { bootstrapEditorIntegration } from './editor/bootstrap';
 import { bootstrapFunctionSearch } from './functions/bootstrap';
+import { bootstrapAiAssistantPromo } from './promo/bootstrap';
 import './editor/indent-guides.css';
 import '../syntax/syntax-highlighting.css';
 import '../themes/themes.css';
@@ -8,3 +9,4 @@ import '../themes/zoho-theme-overrides.css';
 
 void bootstrapEditorIntegration();
 void bootstrapFunctionSearch();
+void bootstrapAiAssistantPromo();